OSCAREMR-6093 Adding patient to waiting list
https://oscaremr.atlassian.net/browse/OSCAREMR-6093

+++ src/main/java/org/oscarehr/common/dao/WaitingListDao.java
@@ -123,11 +123,11 @@ public class WaitingListDao extends AbstractDao<WaitingList> {
 	public Integer getMaxPosition(Integer listId) {
 		Query query = entityManager.createQuery("select max(w.position) from WaitingList w where w.listId = :listId AND w.isHistory = 'N'");
 		query.setParameter("listId", listId);
-		Integer result = (Integer) query.getSingleResult();
+		Long result = (Long) query.getSingleResult();
 		if (result == null) {
 			return 0;
 		} 
-		return result;
+		return result.intValue();
     }
 	
 	@SuppressWarnings("unchecked")
